% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{openreadout_ome_xml}
\alias{openreadout_ome_xml}
\title{OME-XML of a file}
\usage{
openreadout_ome_xml(x)
}
\arguments{
\item{x}{A path or an \code{\link[=openreadout_open]{openreadout_open()}} file.}
}
\value{
A string.
}
\description{
OME-XML (2016-06, \code{MetadataOnly} pixels) describing every image, as an OME-TIFF export would
write it.
}
