% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{openreadout_version}
\alias{openreadout_version}
\title{Version of the bundled OpenReadout library}
\usage{
openreadout_version()
}
\value{
A string such as \code{"0.1.0"}.
}
\description{
Version of the bundled OpenReadout library
}
